Fix: resolve user profile error/modularize auth
