+++ client/package.json
@@ -12,7 +12,6 @@
   "dependencies": {
     "axios": "^1.7.7",
     "formik": "^2.4.6",
-    "framer-motion": "^11.12.0",
     "lucide-react": "^0.454.0",
     "mapbox-gl": "^3.8.0",
     "react": "^18.3.1",
