+++ client/src/components/CategoryIconHeader.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import { useState } from "react";
 import {
   Home,
   Castle,
