+++ client/src/screens/ShowListing.jsx
@@ -2,10 +2,10 @@ import axios from "axios";
 import { useParams, useNavigate } from "react-router-dom";
 import MainScreen from "../components/MainScreen";
 import { useEffect, useState } from "react";
-import { ArrowRightCircle, Bookmark, EditIcon, Trash2Icon } from "lucide-react";
+import { ArrowRightCircle, EditIcon, Trash2Icon } from "lucide-react";
 import LeaveReview from "../components/LeaveReview";
 import AllReviews from "../components/AllReviews";
-import { useAuth } from "../context/AuthContext";
+import { useAuth } from "../hooks/useAuth";
 import ShowMap from "../components/ShowMap";
 import config from "../config";
 import { toast } from "react-toastify";
